Setup service to get new Sudoku
