+++ src/app/grid/grid.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit, Input, Output, EventEmitter, OnChanges, SimpleChanges  } from '@angular/core';
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 
+import { SudokuService } from  '../Service/sudoku.service';
+
+
 @Component({
   selector: 'app-grid',
   templateUrl: './grid.component.html',
@@ -8,7 +11,8 @@ import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 })
 export class GridComponent implements OnInit,OnChanges  {
 
-  solutionData;
+  data;
+
   name = "Sudoku";
   x = "";
   y = "";
@@ -18,18 +22,16 @@ export class GridComponent implements OnInit,OnChanges  {
   isFinished : boolean = false;
 
   ngOnChanges(changes: SimpleChanges) {
-    // Input change handling logic goes here   
-        // alert('inchnage'+this.numberOfCellsDisabled)
-      this.randomizeDisableCells(this.numberOfCellsDisabled);
-
+    this.randomizeDisableCells(this.numberOfCellsDisabled);
   }
 
-  constructor(private modalService: NgbModal) {
-    // alert('Generate now')
-     this.generateSudoku()
+  constructor(  private modalService : NgbModal, 
+                private sudokuService : SudokuService ) 
+  {
 
   }
-   ngOnInit() {
+  ngOnInit() {
+
 
   }
 
@@ -52,114 +54,43 @@ export class GridComponent implements OnInit,OnChanges  {
   //   alert('end is solved'+this.isFinished)
   // }
 
-  changeDifficulty(level){
-
-    this.difficulty.emit(level)
-    // switch (difficulty)
-    // {
-    //   case "easy":
-    //     this.numberOfCellsDisabled = 50;
-    //       break;
-    //   case "medium":
-    //     this.numberOfCellsDisabled = 35;
-    //       break;
-    //   case "advanced": 
-    //     this.numberOfCellsDisabled = 20;
-    //       break;
-    // }
-  }
-
-  removeData(){
-    // for(var j = 0; j < 9; j++) {
-    //   for(var i = 0; i < 9; i++) {
-    //     if(this.data[j][i].s == true)
-    //       this.data[j][i].v = null
-    //   }
-    // }
-  }
-
-
-// START || functions to generate sudoku 
-  generateSudoku(){
-    var data = [1,2,3,4,5,6,7,8,9];
-    var r1,r2,temp ; 
-
-    for(var i = 0; i < 6; i++) {
-      r1 = this.getRandomInt(9); r2 = this.getRandomInt(9)
-      temp = data[r1]; data[r1] = data[r2];data[r2] = temp;
-    }
-
-    this.fillRow(data, 0)    
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 1)
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 2)    
-    this.fillRow(this.shiftArrayByPlaces(data, 1), 3)    
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 4)    
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 5)    
-    this.fillRow(this.shiftArrayByPlaces(data, 1), 6)
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 7)
-    this.fillRow(this.shiftArrayByPlaces(data, 3), 8)
-      
-  }
-
-  //this fills in data to Nth row with passed in data
-  fillRow(rowData, rowNum){
-    for(var j = 0; j < 9; j++)
-      this.data[rowNum][j].v = rowData[j]
-  }
-
-  //Shift array to left by Nth places
-  shiftArrayByPlaces(data,numOfPlaces){
-    for(let i=0; i<numOfPlaces; i++)
-      data = this.shiftArray(data);
-
-    return data;     // alert(JSON.stringify(data));
-  }
-
-    //Shift Array by one place to left
-  shiftArray(data){
-    let temp = data[0];
-
-    for(var i = 0; i < data.length-1; i++) 
-      data[i] = data[i+1]
-    
-    data[data.length-1] = temp
-    
-    return data
-  }
-
-// END || functions to generate sudoku 
-
 // START || functions to disable random cells 
   // Randomly disable cell as per input parameter 
   randomizeDisableCells(noOfCells) {
     var xRan, yRan, counter = 0;
 
-    for(var j = 0; j < 9; j++) {
-      for(var i = 0; i < 9; i++) {
-        this.data[j][i].s == true
-      }
-    }
+    this.data = null;
 
-    while (counter < noOfCells) {
-      xRan = this.getRandomInt(9);
-      yRan = this.getRandomInt(9);  
-      if(this.data[xRan][yRan].s){
-        this.data[xRan][yRan].s = false;
-        counter++;
-      }else{
-        continue;
-      }
-    }
+    this.sudokuService.getSudoku()
+        .subscribe(heroes => this.data = heroes);
 
-    // this.removeData();
+    const solutionData = this.data;
 
     for(var j = 0; j < 9; j++) {
-      for(var i = 0; i < 9; i++) {
-        if(this.data[j][i].s == true)
-          this.data[j][i].v = null
-      }
-    }
+        for(var i = 0; i < 9; i++) {
+          this.data[j][i].s == true
+        }
 
+      }
+        
+      while (counter < noOfCells) {
+        xRan = this.getRandomInt(9);
+        yRan = this.getRandomInt(9);  
+        if(this.data[xRan][yRan].s){
+          this.data[xRan][yRan].s = false;
+          counter++;
+        }else{
+          continue;
+        }
+      }
+
+
+      for(var j = 0; j < 9; j++) {
+        for(var i = 0; i < 9; i++) {
+          if(this.data[j][i].s == true)
+          this.data[j][i].v = " "
+        }
+      }
 
   }
 
@@ -195,57 +126,7 @@ export class GridComponent implements OnInit,OnChanges  {
     this.x = x;
     this.y = y;
     this.modalService.open(comp, { centered: true, size: "sm" });
+  
   }
-
-
-
-
-  data = [
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 1, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 1, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ],
-    [
-      { v: 1, s: true },{ v: 2, s: true },{ v: 3, s: true },
-      { v: 4, s: true },{ v: 5, s: true },{ v: 6, s: true },
-      { v: 7, s: true },{ v: 8, s: true },{ v: 9, s: true }
-    ]
-  ];
 
 }
